+++ pyoupyou/settings/common.py
@@ -123,3 +123,5 @@ DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
 # Process analysis financial starting date.
 # It should only contain the month and day using '<month>-<day>' format.
 FINANCIAL_STARTING_MONTH_DAY = "07-01"
+
+SEND_PLANNING_EMAIL = False